Move AssetsXYZ components
into src/renderer/components/wallets/assets
