+++ src/renderer/components/wallet/assets/AssetDetails.stories.tsx
@@ -4,8 +4,8 @@ import * as RD from '@devexperts/remote-data-ts'
 import { storiesOf } from '@storybook/react'
 import * as O from 'fp-ts/lib/Option'
 
-import { WALLET_ADDRESS_TESTNET } from '../../../shared/mock/address'
-import { ASSETS_TESTNET } from '../../../shared/mock/assets'
+import { WALLET_ADDRESS_TESTNET } from '../../../../shared/mock/address'
+import { ASSETS_TESTNET } from '../../../../shared/mock/assets'
 import AssetDetails from './AssetDetails'
 
 const address = O.some(WALLET_ADDRESS_TESTNET)
