+++ src/renderer/components/wallet/assets/AssetDetails.tsx
@@ -8,16 +8,16 @@ import * as O from 'fp-ts/lib/Option'
 import { useIntl } from 'react-intl'
 import { useHistory } from 'react-router-dom'
 
-import * as AH from '../../helpers/assetHelper'
-import * as walletRoutes from '../../routes/wallet'
-import { TxsRD, LoadTxsProps, SendAction, isSendAction } from '../../services/binance/types'
-import { MAX_PAGINATION_ITEMS } from '../../services/const'
-import { NonEmptyAssetsWithBalance } from '../../services/wallet/types'
-import AssetInfo from '../uielements/assets/AssetInfo'
-import BackLink from '../uielements/backLink'
-import Button, { RefreshButton } from '../uielements/button'
+import * as AH from '../../../helpers/assetHelper'
+import * as walletRoutes from '../../../routes/wallet'
+import { TxsRD, LoadTxsProps, SendAction, isSendAction } from '../../../services/binance/types'
+import { MAX_PAGINATION_ITEMS } from '../../../services/const'
+import { NonEmptyAssetsWithBalance } from '../../../services/wallet/types'
+import AssetInfo from '../../uielements/assets/AssetInfo'
+import BackLink from '../../uielements/backLink'
+import Button, { RefreshButton } from '../../uielements/button'
+import TransactionsTable from '../TransactionsTable'
 import * as Styled from './AssetDetails.style'
-import TransactionsTable from './TransactionsTable'
 
 type SendActionMenuItem = {
   key: SendAction
