+++ src/renderer/components/wallet/assets/AssetsNav.stories.tsx
